Add a 2D grid utility
This was part of a failed attempt at Day 3, Puzzle 2. I attempted a
solution with cellular automata, but that approach got too complicated.
I think I have an alternative that will work more like a
tokenizer/parser, which I'm more optimistic will work. I decided to keep
the grid because it might be useful.